+++ prp/parse/amrfinder.py
@@ -10,7 +10,6 @@ import pandas as pd
 
 from ..models.phenotype import (
     AmrFinderGene,
-    AmrFinderResistanceGene,
     AmrFinderVariant,
     AmrFinderVirulenceGene,
     AMRMethodIndex,
@@ -74,14 +73,14 @@ def _read_result(path: str) -> Tuple[AmrFinderGenes, AmrFinderVariants]:
 
 def _format_gene(
     hit: Dict[str, Any]
-) -> AmrFinderGene | AmrFinderVirulenceGene | AmrFinderResistanceGene:
+) -> AmrFinderGene | AmrFinderVirulenceGene:
     """Format AMRfinder gene."""
     element_type = ElementType(hit["element_type"])
     match element_type:
         case ElementType.VIR:
             gene_type = AmrFinderVirulenceGene
         case ElementType.AMR:
-            gene_type = AmrFinderResistanceGene
+            gene_type = AmrFinderGene
         case _:
             gene_type = AmrFinderGene