+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
  - Set environment variables to avoid interactive prompts in `Dockerfile` by incl. TZ data
+ - Updated `openjdk-17` to `openjdk-25` in `Dockerfile`
 
 ## [1.3.4]